refactor(go): split 818-line structs.go into themed files; extract helpers
Phase 7 of 9. Closes C2 (single-file sprawl + repeated patterns) and
C3 (dead serializeNode dispatcher).

Pure internal refactor. No public API change. JSON output is byte-
identical before and after — verified by 49 Go tests + 99 bun tests
+ smoke all passing unchanged.

Helpers extracted (processor/helpers.go):
- withPos(n, m) — injects pos/end into any serializer's output map.
  Removes ~80 lines of `"pos": nodePos(n.Pos()), "end": nodePos(n.End())`
  boilerplate from ~40 serializers.
- serializeStmts/Words/Assigns/Redirects/ArithmExprs/ArrayElems/
  CaseItems/WordParts/Comments — typed []T → []interface{} mappers.
  Previously open-coded as inline `make([]interface{}, len(xs)); for i, x := range xs { out[i] = serializeX(x) }` ~12 times.

Dead code removed:
- serializeNode (the general-purpose dispatcher) was defined but never
  called. The interface dispatchers (serializeCommand, serializeWordPart,
  serializeArithmExpr, serializeTestExpr, serializeLoop) cover every
  reachable path. Deleted ~75 lines.

Split structs.go (818 lines) into 4 themed files:
- dispatch.go (133): the 5 interface dispatchers
- framing.go (79): File, Stmt, Redirect, Word, Assign, Comment
- commands.go (170): all 16 Command implementers
- expressions.go (231): WordPart implementers + arithmetic/test exprs +
  supporting leaves (ArrayExpr/Elem, WordIter, CStyleLoop, Slice,
  Replace, Expansion)

Largest file is now 231 lines (was 818). Each file has a clear
thematic boundary and matches a section of the audit's node-coverage
table.

Verified: go test 49/49, bun test 99/99, typecheck OK, lint 0
errors / 55 warnings, smoke OK, compiled binary still embeds WASM.
